Rewrite in Jquery
Added an exception for arXiv IDs, which should be resolved in doi2bib (but aren't at the moment, issue opened).

Rewritten the checks in jquery, which handles the exceptions well (link no link).

Switched to Last Names only for search to make it clean.

Included AJAX in the loop. Easier that way and makes better use of async. (Feels more responsive).

Added a check that compares the titles returned by the Crossref query so only provide a link if we actually have valid data. (Sometimes we don't and then clicking the link and copying is a bit annoying.) It says "No DOI found" if the titles do not match. Titles are case insensitive.